refactor(ica/host): moved getter and setters to 'keeper.go'
