--- modules/apps/27-interchain-accounts/host/keeper/params.go
@@ -1,27 +1,0 @@
-package keeper
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/host/types"
-)
-
-// GetParams returns the total set of the host submodule parameters.
-func (k Keeper) GetParams(ctx sdk.Context) types.Params {
-	var params types.Params
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.ParamsKey)
-	if bz == nil {
-		return types.Params{}
-	}
-
-	k.cdc.MustUnmarshal(bz, &params)
-	return params
-}
-
-// SetParams sets the total set of the host submodule parameters.
-func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
-	store := ctx.KVStore(k.storeKey)
-	bz := k.cdc.MustMarshal(&params)
-	store.Set(types.ParamsKey, bz)
-}
